bloodDonation/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-filter==24.2
djangorestframework==3.15.1
docopt==0.6.2
idna==3.7
importlib-metadata==7.1.0
Markdown==3.6
pipreqs==0.4.13
requests==2.31.0
sqlparse==0.5.0
typing-extensions==4.11.0
urllib3==2.2.1
yarg==0.1.9
zipp==3.18.1
